refactor: 채팅 리팩토링
